packages/entrag/pyproject.toml
[project]
name = "entrag"
version = "0.0.0"
description = "EntRAG Benchmarking Tools"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "click>=8.1.8",
    "numpy>=2.2.3",
    "faiss-cpu>=1.10.0",
    "loguru>=0.7.3",
    "pymupdf>=1.25.3",
    "nltk>=3.9.1",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
    "sentence-transformers>=3.4.1",
    "transformers>=4.49.0",
    "markitdown[all]>=0.0.1a5",
    "google-genai>=1.5.0",
    "rapidfuzz>=3.12.2",
    "docling>=2.15.1",
    "pypdf2>=3.0.1",
    "docling-core",
    "matplotlib>=3.10.1",
    "pypdf>=5.4.0",
    "seaborn>=0.13.2",
    "tiktoken>=0.9.0",
    "networkx>=3.4.2",
]

[dependency-groups]
dev = ["entrag-mock-api"]

[tool.poe.tasks]
analyze-dataset = { script = "scripts.analyze-dataset:main" }
build-raw-documents = { script = "scripts.build-raw-documents:main" }
document-ingestion = { script = "scripts.document-ingestion:main" }
compute-classification-metrics = { script = "scripts.compute-classification-metrics:main" }
entrag = { shell = "python -m entrag.main" }

[tool.ruff]
extend = "../../pyproject.toml"
exclude = ["build", "dist"]
include = ["src/**"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
